Initial Commit
Initial commit of the web storm project which includes bootstrap form
with validations using Angular. Validation is dynamic and is handled
using a  custom directive which reads the inputs and apply has Error
only after the input field is blurred.

Future of this project includes service which saves the information
entered in the form.